seperated truncateAmount function and getCurrentDateAndTime function from validation file to their respective files
