+++ src/Core/Helpers/amountUtils.js
@@ -1,0 +1,4 @@
+export const truncateAmount = (amount) => {
+    const amountStr = amount.toString();
+    return amountStr.length > 6 ? `${amountStr.slice(0, 6)}...` : parseFloat(amountStr).toFixed(2);
+  };
