+++ src/Core/Helpers/validations.js
@@ -28,15 +28,4 @@ export const handleAmountBeforeInput = (e) => {
     }
   };
 
-  export const getCurrentDateAndTime = () =>{
-    const formattedDateTime = new Date()
-          .toLocaleDateString("en-GB")
-          .replace(/\//g, "-") + " " + new Date()
-          .toLocaleTimeString("en-GB", { hour: "2-digit", minute: "2-digit", hour12: false });
-    return formattedDateTime
-  }
 
-  export const truncateAmount = (amount) => {
-    const amountStr = amount.toString();
-    return amountStr.length > 6 ? `${amountStr.slice(0, 6)}...` : parseFloat(amountStr).toFixed(2);
-  };
