--- src/Themes/theme.js
@@ -1,22 +1,0 @@
-import { createTheme } from "@mui/material/styles";
-
-const theme = createTheme({
-  // typography: {
-  //   h6: {
-  //     fontWeight: "bold",
-  //     fontSize: "1.15rem",
-  //     color: "#333",
-  //     cursor: "pointer",
-  //   },
-  //   h5: {
-  //       fontSize:"1.9rem",
-  //       marginBottom:"4px"
-  //   },
-  //   h4:{
-        
-  //   }
-  // },
-
-});
-
-export default theme;
